\docType{data}
\name{ManMeter}
\alias{ManMeter}
\title{Example Manufacturer}
\format{A data table with 70176 rows and 3 variables}
\description{
A dataset containing 15-minute smart meter interval data
for an office space in the Sacramento region. Dates range
from Midnight January 1, 2012 to end of day December 31,
2013
}
\details{
The template format for imported meter data is as follows:

\itemize{ \item meter. The identifier of the meter reading.
\item timestamp. The 15-minute time stamp, imported as GMT.
\item kW. The monitored 15-minute demand over the meter. }
}
\keyword{datasets}

