% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runExample}
\alias{runExample}
\title{Run a Shiny Example of EEAR}
\usage{
runExample(example)
}
\arguments{
\item{example}{Name of an available Shiny example}
}
\description{
Launches a Shiny app UI for loading, displaying, and discovering
usage patterns within a customer's consumption behavior.
}

