% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{SchoolMeter}
\alias{SchoolMeter}
\title{Example School Campus}
\format{A data table with 140160 rows and 3 variables}
\description{
A dataset containing 15-minute smart meter interval data for a
  School campus in the Sacramento region. Dates range from
  Midnight July 1, 2012 to end of day June 30, 2013
}
\details{
The template format for imported meter data is as follows:

\itemize{
  \item meter. The identifier of the meter reading.
  \item timestamp. The 15-minute time stamp, imported as GMT.
  \item kW. The monitored 15-minute demand over the meter.
}
}
\keyword{datasets}

