% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{temperatures}
\alias{temperatures}
\title{Interpolated 15-minute temperature values from Sacramento Airport}
\format{A data table with 179928 rows and 2 variables}
\description{
A dataset containing 15-minute interpolated temperatures values
  from the Sacramento Airport weather monitoring station for the
  ranging from the end of 2008 to February 2014.
  The variables are as follows:
}
\details{
\itemize{
  \item timestamp. POSIXct date and time in PST/PDT time zone.
  \item temperature. degrees in F.
  ...
}
}
\keyword{datasets}

