\name{transformations}
\alias{to_degree}
\alias{to_eui}
\alias{to_kwh}
\alias{transformations}
\title{Transform the kW to another measurement (kWh, EUI, per-degree)}
\usage{
to_kwh(x)

to_eui(x)

to_degree(x)
}
\arguments{
  \item{x}{a table object.}
}
\description{
Transform the kW to another measurement (kWh, EUI,
per-degree)
}

